requirements.txt
flask>=3.0.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
joblib>=1.2.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
PyPDF2>=3.0.0
tqdm>=4.64.0
requests>=2.31.0
python-dotenv>=1.0.0
